refactor notifications component
