+++ src/components/common/BaseLayout.jsx
@@ -47,7 +47,7 @@ const BaseLayoutComponent = ({
         <AuthorizationHOCUnit exact path='/article/new' component={CreateArticlePage} />
         <Route exact path='/article/:slug' component={Article} />
         <Route path='/auth/social' component={AuthSocialComponent} />
-        <AuthorizationHOCUnit exact path='/profiles/notification' component={NotificationPage} />
+        <AuthorizationHOCUnit exact path='/profile/notifications' component={NotificationPage} />
         <AuthorizationHOCUnit exact path='/profile' component={Profile} />
         <AuthorizationHOCUnit exact path='/editProfile' component={EditProfile} />
         <Route path='/signup' component={SignUpDefault} />
